requirements.txt
opencv-python
mediapipe

torch
torchaudio
transformers
torchvision
presidio-analyzer
presidio-anonymizer

openai-whisper
pydub

moviepy